Starting with TUI
This is a text based user input solution I have been working on for the time being while I investigate further into Python's Tkinter library. It's still a work in progress but feel free to tinker with it (pun intended). It should technically work, however be weary that it is not fully functional and may break under certain file conditions.